refactor: 형식 수정
- move package DisplayHandler
- change OutputFormatter string concat method
- change access modifier of constant
